pyproject.toml
[project]
name = "gemini-voice-input"
version = "0.2.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "evdev>=1.9.3",
    "google-genai>=1.64.0",
    "keyboard>=0.13.5",
    "numpy>=2.4.2",
    "scipy>=1.17.0",
    "sounddevice>=0.5.5",
]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "ruff>=0.15.2",
]
